feat(clerk-js): Use TypeScript for sandbox
