+++ packages/clerk-js/sandbox/template.html
@@ -133,7 +133,7 @@
           viewBox="0 0 62 18"
           fill="none"
           aria-hidden="true"
-          class="h-[1.125rem] text-gray-950 dark:text-white"
+          class="h-[1.125rem] text-gray-950"
         >
           <ellipse
             cx="8.99999"
